Function References

Function References

JuMag.FDMesh — Type.

FDMesh(;dx=1e-9, dy=1e-9, dz=1e-9, nx=1, ny=1, nz=1, pbc="open")

Create a FDMesh for given parameters. pbc could be any combination of "x", "y" and "z".

JuMag.Sim — Function.

Sim(mesh::Mesh; driver="LLG", name="dyn", integrator="Dopri5")

Create a simulation instance for given mesh.

JuMag.set_Ms — Function.

set_Ms(sim::MicroSim, init::Any) Set the magnetization of the simulation eara, which can be a constant or a function with parameters of (i,j,k,dx,dy,dz)

Missing docstring.

Missing docstring for init_m0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_exch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_dmi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_dmi_interfacial. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_demag. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_zeeman. Check Documenter's build log for details.

JuMag.update_zeeman — Function.

updatezeeman(z::Zeeman, H0::Tuple) If you have addzeeman, and want to change the field, use

Missing docstring.

Missing docstring for relax. Check Documenter's build log for details.