Function References

Function References

JuMag.FDMesh — Type.

FDMesh(;dx=1e-9, dy=1e-9, dz=1e-9, nx=1, ny=1, nz=1, pbc="open")

Create a FDMesh for given parameters. pbc could be any combination of "x", "y" and "z".

JuMag.Sim — Function.

Sim(mesh::Mesh; driver="LLG", name="dyn", integrator="Dopri5")

Create a simulation instance for given mesh.

JuMag.set_Ms — Function.

set_Ms(sim::MicroSim, Ms::NumberOrArrayOrFunction)

Set the saturation magnetization of the studied system. As can be guessed from the type union NumberOrArrayOrFunction, Ms could be a number or an array or a function. If Ms is an array, its length should be equal to the size of the system. If Ms is a function, it should take six parameters in the form (i,j,k,dx,dy,dz) where i,j,k are the indices of the mesh cells and dx,dy,dz are the cellsizes. For example,

   set_Ms(sim, 8.6e5)

or

function circular_Ms(i,j,k,dx,dy,dz)
    if (i-50.5)^2 + (j-50.5)^2 <= 50^2
        return 8.6e5
    end
    return 0.0
end
set_Ms(sim, circular_Ms)

JuMag.init_m0 — Function.

init_m0(sim::MicroSim, m0::TupleOrArrayOrFunction; norm=true)

Set the initial magnetization of the system, the input m0 could be a Tuple with length 3 or an array with length 3*N where N is the system size or a function with six parameters (i,j,k,dx,dy,dz).

Missing docstring.

Missing docstring for add_exch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_dmi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_dmi_interfacial. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_demag. Check Documenter's build log for details.

JuMag.add_zeeman — Function.

add_zeeman(sim::AbstractSim, H0::TupleOrArrayOrFunction; name="zeeman")

Add a static Zeeman energy to the simulation.

add_zeeman(sim::AbstractSim, H0::TupleOrArrayOrFunction, ft::Function; name="timezeeman")

Add a time varying zeeman to system. The input H0 could be

A tuple with length 3, for example, (0,0,1e2).

An array with length 3N where N is the number of total spins.

A function with six parameters (i,j,k,dx,dy,dz) and return a tuple, for example,
 function linear_Hz(i, j, k, dx, dy, dz)
    nx = 100
    Hz = i/nx*1000
    return (0, 0, Hz)
 end

The input ft is a function of time t and its return value should be a tuple. For example,

  function time_fun(t)
    return (sin(t), cos(t), 0)
  end

JuMag.update_zeeman — Function.

update_zeeman(z::Zeeman, H0::Tuple)

Set the external field of Zeeman z to H0 where H0 is a tuple. For example,

   ze = add_zeeman(sim, (0,0,0)) #create a zeeman energy with field (0,0,0) A/m
   update_zeeman(ze, (0,0,1e5)) #change the field to (0,0,1e5) A/m

Missing docstring.

Missing docstring for relax. Check Documenter's build log for details.