Function References

Function References

JuMag.FDMesh — Type.

FDMesh(;dx=1e-9, dy=1e-9, dz=1e-9, nx=1, ny=1, nz=1, pbc="open")

Create a FDMesh for given parameters. pbc could be any combination of "x", "y" and "z".

JuMag.Sim — Function.

Sim(mesh::Mesh; driver="LLG", name="dyn", integrator="Dopri5")

Create a simulation instance for given mesh.

JuMag.NumberOrArrayOrFunction — Constant.

NumberOrArrayOrFunction

In Micromagnetics, typical parameters such as saturation magnetization Ms and exchange stiffness constant A are constant. However, there are many cases that a spatial Ms is needed. For instance, if the simulated system is a circular disk the Ms in the corners should be set to zero. If the simulated system contains mutiple materials, the exchange constant A should be spatial as well. The Union NumberOrArrayOrFunction is designed to deal with such situations. As indicated from its name, it means that the input parameter could be a number or an array or a function:

Number: should be Real.

Array: the length of the array should be N where N is the total spin number of the system.

Function: the parameter of the function should be (i,j,k,dx,dy,dz) where i,j,k is the cell index and dx,dy,dz is the cellsize. The return value of the function should be a real number. For example,

function circular_Ms(i,j,k,dx,dy,dz)
    if (i-50.5)^2 + (j-50.5)^2 <= 50^2
        return 8.6e5
    end
    return 0.0
end

JuMag.TupleOrArrayOrFunction — Constant.

TupleOrArrayOrFunction

Similar to NumberOrArrayOrFunction, TupleOrArrayOrFunction means that the input parameter could be a tuple or an array or a function:

Tuple: should be Real with length 3. For example, (0,0,1e5).

Array: the length of the array should be 3N where N is the total spin number of the system.

Function: the parameter of the function should be (i,j,k,dx,dy,dz) and the return value should be a tuple with length 3. For example,
function uniform_m0(i,j,k,dx,dy,dz)
    return (0,0,1)
end

JuMag.set_Ms — Function.

set_Ms(sim::MicroSim, Ms::NumberOrArrayOrFunction)

Set the saturation magnetization Ms of the studied system. For example,

   set_Ms(sim, 8.6e5)

or

function circular_Ms(i,j,k,dx,dy,dz)
    if (i-50.5)^2 + (j-50.5)^2 <= 50^2
        return 8.6e5
    end
    return 0.0
end
set_Ms(sim, circular_Ms)

JuMag.init_m0 — Function.

init_m0(sim::MicroSim, m0::TupleOrArrayOrFunction; norm=true)

Set the initial magnetization of the system. If norm=false the magnetization array will be not normalised. Examples:

   init_m0(sim, (1,1,1))

or

   init_m0(sim, (1,1,1), norm=false)

or

   function uniform_m0(i,j,k,dx,dy,dz)
       return (0,0,1)
   end
   init_m0(sim, uniform_m0)

JuMag.add_exch — Function.

add_exch(sim::AbstractSim, A::NumberOrArrayOrFunction; name="exch")

Add exchange energy to the system.

JuMag.add_anis — Function.

add_anis(sim::AbstractSim, Ku::NumberOrArrayOrFunction; axis=(0,0,1), name="anis")

Add Anisotropy to the system, where the energy density is given by

\[E_\mathrm{anis} = - K_{u} (\vec{m} \cdot \hat{u})^2\]

JuMag.add_dmi — Function.

add_dmi(sim::AbstractSim, D::Tuple{Real, Real, Real}; name="dmi")

Add DMI to the system. Example:

   add_dmi(sim, (1e-3, 1e-3, 0))

add_dmi(sim::AbstractSim, D::Real; name="dmi", type="bulk")

Add DMI to the system. type could be "bulk" or "interfacial" Examples:

   add_dmi(sim, 1e-3, type="interfacial")

or

   add_dmi(sim, 1e-3, type="bulk")

JuMag.add_demag — Function.

add_demag(sim::MicroSim; name="demag", Nx=0, Ny=0, Nz=0)

Add Demag to the system. Nx, Ny and Nz can be used to describe the macro boundary conditions which means that the given mesh is repeated 2Nx+1, 2Ny+1 and2Nz+1times inx,yandz` direction, respectively.

JuMag.add_zeeman — Function.

add_zeeman(sim::AbstractSim, H0::TupleOrArrayOrFunction; name="zeeman")

Add a static Zeeman energy to the simulation.

add_zeeman(sim::AbstractSim, H0::TupleOrArrayOrFunction, ft::Function; name="timezeeman")

Add a time varying zeeman to system.

The input ft is a function of time t and its return value should be a tuple with length 3.

Example:

  function time_fun(t)
    w = 2*pi*2.0e9
    return (sin(w*t), cos(w*t), 0)
  end

  function spatial_H(i, j, k, dx, dy, dz)
    H = 1e3
    if i<=2
        return (H, H, 0)
    end
    return (0, 0, 0)
  end

  add_zeeman(sim, spatial_H, time_fun)

JuMag.update_zeeman — Function.

update_zeeman(sim::AbstractSim, H0::Tuple; name="zeeman")

Set the external field of Zeeman z to H0 where H0 is TupleOrArrayOrFunction. For example,

   add_zeeman(sim, (0,0,0)) #create a zeeman energy with field (0,0,0) A/m
   update_zeeman(sim, (0,0,1e5)) #change the field to (0,0,1e5) A/m

JuMag.update_anis — Function.

update_anis(sim::MicroSim, Ku::NumberOrArrayOrFunction; name = "anis")

Replace the value of Anisotropy with Ku Example:

julia mesh = FDMesh(nx=200, ny=200, nz=12, dx=5e-9, dy=5e-9, dz=5e-9) sim = Sim(mesh) add_anis(sim,3e4,axis = (0,0,1)) update_anis(sim, 5e4)`

JuMag.relax — Function.

relax(sim::AbstractSim; maxsteps=10000, stopping_dmdt=0.01, stopping_torque=0.1, save_m_every = 10, save_ovf_every=-1, ovf_format = "binary", ovf_folder="ovfs", save_vtk_every=-1, vtk_folder="vtks",fields::Array{String, 1} = String[])

Relax the system using LLG or SD driver. The stop condition is determined by either stopping_dmdt(for LLG) or stopping_torque(for SD). Spins can be stored in ovfs or vtks. ovf format can be chosen in "binary"(float64),"binary8"(float64), "binary4"(float32), "text"

Fields can be stored in vtks by:

relax(sim,save_vtk_every = 10,fields = ["demag","exch","anis"])

Missing docstring.

Missing docstring for save_m. Check Documenter's build log for details.